Move options map into xlator package
...keeps code simpler.

The xlators map and options map are now private. xlator.Find and
xlator.FindOption helper functions are used to search the maps.
