+++ glusterd2/xlator/global.go
@@ -9,11 +9,15 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-// Xlators is map of all available xlators, indexed by xlator-id
-// Other packages can directly import this.
-var Xlators map[string]*Xlator
+var (
+	// xlMap is a map of all available xlators, indexed by xlator-id
+	xlMap map[string]*Xlator
+	// options is a map of all available options indexed by
+	// <xlator-id>.<option-key> for all keys of an option
+	optMap map[string]*options.Option
+)
 
-// Load initializes the global variable xlator.Xlators and options.Options
+// Load load all available xlators and intializes the xlators and options maps
 func Load() (err error) {
 
 	defer func() {
@@ -28,22 +32,22 @@ func Load() (err error) {
 	if err != nil {
 		return
 	}
-	Xlators = xls
+	xlMap = xls
 
-	// Also prepare the option.Options map
 	loadOptions()
 
 	return
 }
 
 // loadOptions loads all available options into the options.Options map,
-// indexed as <xlator.ID>.<option.Key> for all available option keys
+// indexed as <xlator-id>.<option-key> for all available option keys
 func loadOptions() {
-	for _, xl := range Xlators {
+	optMap = make(map[string]*options.Option)
+	for _, xl := range xlMap {
 		for _, opt := range xl.Options {
 			for _, k := range opt.Key {
 				k := xl.ID + "." + k
-				options.Options[k] = opt
+				optMap[k] = opt
 			}
 		}
 	}
